Bug: maintaining row selection was triggering false selection change
Since capture/restore the selected row, a selection change was triggering
every draw on both Active Plots and Archived Plots.  This was sometimes
squashing client.logPlotId and making the log not update.

As a bonus because I needed it for debugging, we now include the current
plot id in the log title.